requirements.txt
# PostgreSQL (Supabase)
psycopg2-binary==2.9.9

# Cassandra
cassandra-driver==3.29.1

# MongoDB
pymongo==4.6.1

# Neo4j
neo4j==5.16.0

# Redis
redis==5.0.1

# Utilidades
python-dotenv==1.0.0
